Fix Forward() livelock once command_id exceeds 65535
TransactionExecution::Forward() drives the tx state machine in a
`while (true)` loop and breaks out (yielding the processor) when a full
iteration makes no progress, detected via
`curr_op == prev_op && cmd_id == CommandId()`.

`cmd_id` was declared `uint16_t` while `CommandId()` returns `uint32_t`.
Once a transaction's command_id reaches 65536, the snapshot
`cmd_id = CommandId()` is truncated to its low 16 bits, so a `uint16_t`
value (always <= 65535) can never equal a `CommandId()` that is >= 65536.
The no-progress break therefore never fires past that point: instead of
yielding when an operation is blocked on an async cc request, Forward()
spins in place.

Impact: a single transaction issuing more than 65536 commands (e.g. one
delete command removing ~100k documents by _id) livelocks if any of its
reads must block on the cc processor after crossing the boundary. Because
Forward() never yields, the pending ReadCc is never serviced, the read
result stays unfinished (ref_cnt=0, term=-1), and the worker spins at
100% CPU with command_id frozen at exactly 65536. The node is effectively
wedged and does not recover (its shutdown path also blocks).

Fix: declare `cmd_id` as `uint32_t` to match `CommandId()`.

Reproduced and verified against EloqDoc with the MongoDB jstest
core/batch_write_command_delete.js (single delete of 100k _id docs
preceded by a remove({}) that forces async reads): the transaction hung
deterministically before this change and completes correctly after
(delete count and post-delete count assertions pass).

Note: the same uint16_t truncation of command_id exists in the liveness
tracker (CheckWaitingTxs / EnlistWaitingTx / TxProgress::cmd_id_ in
tx_service.h) and in remote Cc request handling
(remote/remote_cc_request.h). Those are not the trigger for this specific
livelock and are left for a follow-up.


bump eloqstore

bump eloqstore

bump eloqstore

bump eloqstore

+++ tx_service/src/tx_execution.cpp
@@ -457,7 +457,7 @@ TxErrorCode TransactionExecution::ConvertCcError(CcErrorCode error)
 TxmStatus TransactionExecution::Forward()
 {
     TransactionOperation *prev_op = nullptr;
-    uint16_t cmd_id = 0;
+    uint32_t cmd_id = 0;
 
     if (!AcquireExclusiveForwardLatch())
     {
